requirements.txt
uvicorn~=0.15.0
requests~=2.26.0
fastapi~=0.70.0
giturlparse~=0.10.0
python-decouple~=3.5